requirements.txt
Flask==2.1.2
gevent==21.12.0
numpy==1.22.3
tensorflow-cpu==2.8.0
Werkzeug==2.1.1
gunicorn==20.1.0